requirements.txt
python3
py3-pip
py3-flask
py3-email-validator
py-gunicorn